feat: support tailwindcss v3.1.x and v3.2.y (#21)
### Summary

Support Tailwind CSS v3.1 and Tailwind CSS v3.2.

---

### Details

Since Tailwind CSS support using ESM configuration in V3.3, we need to
use CJS configuration before that.

As mentioned in
https://github.com/rspack-contrib/rsbuild-plugin-tailwindcss/issues/7#issuecomment-2472754144,
we use `readFile` + `require.resolve` to get the version of
`tailwindcss/package.json`.

- If `satisfies(version, ^3.3.0)`, we will generate ESM configuration to
support both ESM and CJS.
- Else, we generate CJS configuration.

---

### Test plan

We setup two new entries in the testing matrix

- tailwindcss v3.1.0 with Ubuntu
- tailwindcss v3.1.0 with Windows

As you can see, the test is failing at
[0e05c6a](https://github.com/rspack-contrib/rsbuild-plugin-tailwindcss/pull/21/commits/0e05c6a1d04d47f40c580bf3118419a8c3b69bdf).

And after the fix is landed in
[e783415](https://github.com/rspack-contrib/rsbuild-plugin-tailwindcss/pull/21/commits/e783415dcee20bd949400d68a4995b98515b8ed8),
the test for old tailwindcss is passing.

---

### Related links

close: #18 

- Tailwind CSS v3.3:
  - https://github.com/tailwindlabs/tailwindcss/releases/tag/v3.3.0
  - https://github.com/tailwindlabs/tailwindcss/pull/10785
- Tailwind CSS v3.2
  - https://github.com/tailwindlabs/tailwindcss/releases/tag/v3.2.0
- Tailwind CSS v3.1
  - https://github.com/tailwindlabs/tailwindcss/releases/tag/v3.1.0